requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.45
psycopg[binary]==3.3.2
pydantic-settings==2.12.0
python-dotenv==1.2.1
alembic